+++ app/src/constants/activities.ts
@@ -50,21 +50,3 @@ export enum FormValidationStatus {
   INVALID = 'Invalid',
   VALID = 'Valid'
 }
-
-export interface IActivity {
-  _id: string,
-  activityId: string,
-  docType: string,
-  activityType: ActivityType,
-  activitySubtype: ActivitySubtype,
-  status: string,
-  sync: {
-    ready: boolean,
-    status: string,
-    error: string
-  },
-  dateCreated: Date,
-  dateUpdated: Date,
-  formData: any,
-  formStatus: string
-}
